requirements-dev.txt
aws-cdk-lib==2.114.1
constructs>=10.0.0,<11.0.0
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0